+++ src/Components/Stats.svelte
@@ -1,58 +1,258 @@
 <script lang="ts">
-  import { closeImpliedLinks, complement, copy } from "src/sharedFunctions";
-
+  import type { Edge, Graph } from "graphlib";
+  import { sum } from "lodash";
+  import type { Directions } from "src/interfaces";
+  import { DIRECTIONS } from "src/constants";
   import type BreadcrumbsPlugin from "src/main";
+  import {
+    closeImpliedLinks,
+    complement,
+    copy,
+    hierToStr,
+    mergeGs,
+  } from "src/sharedFunctions";
 
   export let plugin: BreadcrumbsPlugin;
 
   const { settings } = plugin;
+  const { userHierarchies } = settings;
   const separator = settings.trailSeperator;
 
-  const { gParents, gSiblings, gChildren } = plugin.currGraphs;
-  const allRG = [gParents, gSiblings, gChildren];
+  const hierGs = plugin.currGraphs;
 
-  const [pRNodes, sRNodes, cRNodes] = allRG.map((g) => g.nodes());
+  function fillInInfo(
+    dir: Directions,
+    gType: string,
+    hierData: {
+      [dir in Directions]: {
+        [graphs: string]: {
+          graph?: Graph;
+          nodes: string[];
+          nodesStr: string;
+          edges: Edge[];
+          edgesStr: string;
+        };
+      };
+    }
+  ) {
+    const g = hierData[dir][gType].graph;
 
-  const [pRNodesStr, sRNodesStr, cRNodesStr] = [
-    pRNodes.join("\n"),
-    sRNodes.join("\n"),
-    cRNodes.join("\n"),
-  ];
+    hierData[dir][gType].nodes = hierData[dir][gType].graph.nodes();
+    hierData[dir][gType].nodesStr = hierData[dir][gType].nodes.join("\n");
 
-  const [pREdges, sREdges, cREdges] = allRG.map((g) => g.edges());
+    hierData[dir][gType].edges = hierData[dir][gType].graph.edges();
+    hierData[dir][gType].edgesStr = hierData[dir][gType].edges
+      .map((e) => `${e.v} → ${e.w}`)
+      .join("\n");
+  }
 
-  const [pREdgesStr, sREdgesStr, cREdgesStr] = [
-    pREdges.map((e) => `${e.v} → ${e.w}`).join("\n"),
-    sREdges.map((e) => `${e.v} → ${e.w}`).join("\n"),
-    cREdges.map((e) => `${e.v} → ${e.w}`).join("\n"),
-  ];
+  const data = hierGs.map((hier) => {
+    const hierData: {
+      [dir in Directions]: {
+        [graphs: string]: {
+          graph?: Graph;
+          nodes: string[];
+          nodesStr: string;
+          edges: Edge[];
+          edgesStr: string;
+        };
+      };
+    } = {
+      up: { Merged: {}, Closed: {}, Implied: {} },
+      same: { Merged: {}, Closed: {}, Implied: {} },
+      down: { Merged: {}, Closed: {}, Implied: {} },
+    };
+    DIRECTIONS.forEach((dir) => {
+      // Merged Graphs
+      /// Smoosh all fieldGs from one dir into a merged graph for that direction as a whole
+      hierData[dir].Merged.graph = mergeGs(...Object.values(hier[dir]));
+      fillInInfo(dir, "Merged", hierData);
 
-  const [closedP, closedS, closedC] = [
-    closeImpliedLinks(gParents, gChildren),
-    closeImpliedLinks(gSiblings, gSiblings),
-    closeImpliedLinks(gChildren, gParents),
-  ];
-  const allG = [closedP, closedS, closedC];
+      // Closed graphs
+      if (dir !== "same") {
+        hierData[dir].Closed.graph = closeImpliedLinks(
+          hierData[dir].Merged.graph,
+          mergeGs(...Object.values(hier[dir === "up" ? "down" : "up"]))
+        );
+      } else {
+        hierData[dir].Closed.graph = closeImpliedLinks(
+          hierData[dir].Merged.graph,
+          hierData[dir].Merged.graph
+        );
+      }
+      fillInInfo(dir, "Closed", hierData);
+    });
 
-  const [pANodes, sANodes, cANodes] = allG.map((g) => g.nodes());
-  const [pAEdges, sAEdges, cAEdges] = allG.map((g) => g.edges());
+    return hierData;
+  });
 
-  const [pIEdgesStr, sIEdgesStr, cIEdgesStr] = [
-    complement(
-      pAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
-      pREdges.map((e) => `${e.v} ${separator} ${e.w}`)
-    ).join("\n"),
-    complement(
-      sAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
-      sREdges.map((e) => `${e.v} ${separator} ${e.w}`)
-    ).join("\n"),
-    complement(
-      cAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
-      cREdges.map((e) => `${e.v} ${separator} ${e.w}`)
-    ).join("\n"),
-  ];
+  console.log({ data });
+
+  let hierStrs: string[] = userHierarchies.map(hierToStr);
+  console.log({ hierStrs });
+
+  const allUpGs = hierGs
+    .map((hier) => hier.up)
+    .map((fields) => Object.values(fields));
+
+  console.log({ allUpGs });
+
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+  //
+
+  //   const allR = Object.values(graphs);
+
+  //   const allRNodes = allR.map((g) => g.nodes());
+
+  //   const allRNodesStr = allRNodes.map((rNodes) => rNodes.join("\n"));
+
+  //   const allREdges = allR.map((g) => g.edges());
+
+  //   const allREdgesStr = allREdges.map((edges) =>
+  //     edges.map((e) => `${e.v} → ${e.w}`).join("\n")
+  //   );
+
+  //   const [closedP, closedS, closedC] = [
+  //     closeImpliedLinks(gParents, gChildren),
+  //     closeImpliedLinks(gSiblings, gSiblings),
+  //     closeImpliedLinks(gChildren, gParents),
+  //   ];
+  //   const allG = [closedP, closedS, closedC];
+
+  //   const [pANodes, sANodes, cANodes] = allG.map((g) => g.nodes());
+  //   const [pAEdges, sAEdges, cAEdges] = allG.map((g) => g.edges());
+
+  //   const [pIEdgesStr, sIEdgesStr, cIEdgesStr] = [
+  //     complement(
+  //       pAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
+  //       pREdges.map((e) => `${e.v} ${separator} ${e.w}`)
+  //     ).join("\n"),
+  //     complement(
+  //       sAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
+  //       sREdges.map((e) => `${e.v} ${separator} ${e.w}`)
+  //     ).join("\n"),
+  //     complement(
+  //       cAEdges.map((e) => `${e.v} ${separator} ${e.w}`),
+  //       cREdges.map((e) => `${e.v} ${separator} ${e.w}`)
+  //     ).join("\n"),
+  //   ];
 </script>
 
+<table>
+  <thead>
+    <tr>
+      <th scope="col">Hierarchy</th>
+      <th scope="col" colspan="5">Count</th>
+    </tr>
+  </thead>
+
+  <tr>
+    <td />
+    <td>Measure</td>
+    <td>Up</td>
+    <td>Same</td>
+    <td>Down</td>
+    <td>Total</td>
+  </tr>
+
+  {#each userHierarchies as hier, i}
+    <tr>
+      <td rowspan="2">
+        {hierStrs[i]}
+      </td>
+      <td>Real Nodes</td>
+      {#each ["up", "same", "down"] as dir}
+        <td
+          aria-label={data[i][dir].Merged.nodesStr}
+          on:click={async () => await copy(data[i][dir].Merged.nodesStr)}
+        >
+          {data[i][dir].Merged.nodes.length}
+        </td>
+      {/each}
+      <td
+        aria-label={[
+          data[i].up.Merged.nodesStr,
+          data[i].same.Merged.nodesStr,
+          data[i].down.Merged.nodesStr,
+        ].join("\n")}
+        on:click={async () =>
+          await copy(
+            [
+              data[i].up.Merged.nodesStr,
+              data[i].same.Merged.nodesStr,
+              data[i].down.Merged.nodesStr,
+            ].join("\n")
+          )}
+      >
+        {[
+          ...data[i].up.Merged.nodes,
+          ...data[i].same.Merged.nodes,
+          ...data[i].down.Merged.nodes,
+        ].length}
+      </td>
+    </tr>
+    <tr>
+      <td>Edges</td>
+      {#each ["up", "same", "down"] as dir}
+        <td
+          aria-label={data[i][dir].Merged.edgesStr}
+          on:click={async () => await copy(data[i][dir].Merged.edgesStr)}
+        >
+          {data[i][dir].Merged.edges.length}
+        </td>
+      {/each}
+      <td
+        aria-label={[
+          data[i].up.Merged.edgesStr,
+          data[i].same.Merged.edgesStr,
+          data[i].down.Merged.edgesStr,
+        ].join("\n")}
+        on:click={async () =>
+          await copy(
+            [
+              data[i].up.Merged.edgesStr,
+              data[i].same.Merged.edgesStr,
+              data[i].down.Merged.edgesStr,
+            ].join("\n")
+          )}
+      >
+        {[
+          ...data[i].up.Merged.edges,
+          ...data[i].same.Merged.edges,
+          ...data[i].down.Merged.edges,
+        ].length}
+      </td>
+    </tr>
+  {/each}
+</table>
+
+<!-- 
 <table>
   <thead>
     <tr>
@@ -105,8 +305,7 @@
     >
     <td>{pREdges.length + sREdges.length + cREdges.length}</td>
   </tr>
-</table>
-
+</table> -->
 <style>
   table {
   }
@@ -118,5 +317,10 @@
   th {
     padding: 3px;
     border: 1px solid var(--text-accent);
+    white-space: pre-line;
+  }
+
+  td.filler {
+    opacity: 0;
   }
 </style>