+++ src/interfaces.ts
@@ -2,9 +2,7 @@ import type { Graph } from "graphlib";
 import type { FrontMatterCache, Pos, TFile } from "obsidian";
 
 export interface BreadcrumbsSettings {
-  parentFieldName: string;
-  siblingFieldName: string;
-  childFieldName: string;
+  userHierarchies: userHierarchy[];
   indexNote: string[];
   refreshIntervalTime: number;
   defaultView: boolean;
@@ -43,6 +41,13 @@ export interface dvFrontmatterCache {
     | TFile;
 }
 
+export type Directions = "up" | "same" | "down";
+export interface userHierarchy {
+  up: string[];
+  same: string[];
+  down: string[];
+}
+
 export interface dvLink {
   display: any;
   embded: boolean;
@@ -64,6 +69,8 @@ export interface neighbourObj {
   siblings: string[];
   children: string[];
 }
+
+export type relObj = { [key: string]: string[] } | { current: TFile };
 
 export interface ParentObj {
   current: TFile;